LPS-24086 Allow plugins to use CMISearchQueryBuilder
