--- portal-impl/src/com/liferay/portal/repository/cmis/search/CMISConjunction.java
@@ -1,58 +1,0 @@
-/**
- * Copyright (c) 2000-2011 Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portal.repository.cmis.search;
-
-import com.liferay.portal.kernel.util.StringBundler;
-import com.liferay.portal.kernel.util.StringPool;
-
-import java.util.List;
-
-/**
- * @author Mika Koivisto
- */
-public class CMISConjunction extends CMISJunction {
-
-	@Override
-	public String toQueryFragment() {
-		if (isEmpty()) {
-			return StringPool.BLANK;
-		}
-
-		List<CMISCriterion> cmisCriterions = list();
-
-		StringBundler sb = new StringBundler(cmisCriterions.size() * 2 - 1);
-
-		if (cmisCriterions.size() > 1) {
-			sb.append("(");
-		}
-
-		for (int i = 0; i < cmisCriterions.size(); i++) {
-			CMISCriterion cmisCriterion = cmisCriterions.get(i);
-
-			if (i != 0) {
-				sb.append(" AND ");
-			}
-
-			sb.append(cmisCriterion.toQueryFragment());
-		}
-
-		if (cmisCriterions.size() > 1) {
-			sb.append(")");
-		}
-
-		return sb.toString();
-	}
-
-}
